tweek params for preparing matrix for clustering. Note: with too much data R clustering will crash linux :(

+++ tweetTransform.py
@@ -249,7 +249,7 @@ def tfidfData(summaryStemmedTweets, summaryTfidfTweets, summaryStopWords,
 	    lib.TFIDF_UseStopWordList_Run(self.obj, stemFileIn, tfidfFileOut, stopWordFileInOut, dictionaryFileOut)
 
     tfidf = TFIDF()
-    tfidf.preRun(summaryStemmedTweets, summaryTfidfTweets, threshold, summaryStopWords)
+    tfidf.preRun(summaryStemmedTweets, summaryTfidfTweets, c_double(threshold), summaryStopWords)
 
     tfidf = TFIDF()
     tfidf.run(summaryStemmedTweets, summaryTfidfTweets, summaryStopWords, dictionaryFile)
